requirements.txt
# Core dependencies for household resilience monitoring
requests>=2.31.0
pyyaml>=6.0
schedule>=1.2.0
pandas>=2.0.0
matplotlib>=3.7.0
python-dateutil>=2.8.2

# Notification dependencies
twilio>=8.5.0
plyer>=2.1.0  # Desktop notifications

# Web dashboard
flask>=3.0.0
gunicorn>=20.1.0

# Data processing
beautifulsoup4>=4.12.0  # Web scraping
feedparser>=6.0.10      # RSS feeds
python-dotenv>=1.0.0    # Environment variables

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0           # Code formatting
flake8>=6.0.0          # Linting

# Optional: Advanced features
# selenium>=4.11.0      # For dynamic web content
# asyncio>=3.4.3        # Async operations
# sqlalchemy>=2.0.0     # If database needed later
